huge discovery made which makes this plugin much simpler
The PlayerPostThink hook is called before player weapons shoot, and PlayerUse is called after. By rewinding monsters in PostThink, the default sven code (or custom weapon script code) can be used as is. No more hacking CVars and weapon entities to override their shoot code! The "unfixable" bugs are now fixed, for free.

The hook call cycle is per-player, so players can shoot in the same server frame without interfering with the rewind logic for each other. Monsters can be tagged with a special value before a player shoots so you know afterwards if it was hit or not.

Example hook call order for players shooting in the same server frame:
1) PostThink 		(player 1)
2) WeaponPrimaryAttack 	(player 1)
3) PlayerUse 		(player 1)
4) PostThink 		(player 2)
5) WeaponPrimaryAttack 	(player 2)
6) PlayerUse 		(player 2)