fix compensation bug with tanks
also experimented with compensating moving platforms. Abusing SVC_DELTAPACKETENTITIES probably can't work, although I reverse-engineered a lot of it and managed to not crash the game sending a custom one. The problem is the packet index will always be out of order or duplicated which triggers CL_FlushEntityPacket which resyncs the entities to the client. This means I can't edit the position of an entity for only one client.

In order to show a predicted platform for every player, 32 copies would need to be made and networked to every player. This will cause high network usage and probably desyncs. It's also weird transitioning from a predicted platform to a non-predicted one. Especially if you have to jump on the same one to progress (fm_fun).